requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-nested-routers==0.92.5
flake8==3.8.4
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
